+++ src/insert/spacefinder/spacefinder.ts
@@ -7,7 +7,13 @@ import fastdom from 'utils/fastdom-promise';
 import { init as initSpacefinderDebugger } from './spacefinder-debug-tools';
 
 type RuleSpacing = {
+	/**
+	 * Minimum distance between an ad and the element, where the element is above the ad
+	 */
 	minAboveSlot: number;
+	/**
+	 * Minimum distance between an ad and the element, where the element is below the ad
+	 */
 	minBelowSlot: number;
 	bypassMinBelow?: string;
 };
@@ -28,26 +34,49 @@ type SpacefinderItem = {
 type SpacefinderRules = {
 	bodySelector: string;
 	body?: HTMLElement | Document;
-	// selector(s) for the elements that we want to allow inserting ads above
+	/**
+	 * Selector(s) for the elements that we want to allow inserting ads above
+	 */
 	candidateSelector: string | string[];
-	// minimum from slot to top of page
+	/**
+	 * Minimum distance from slot to top of page
+	 */
 	absoluteMinAbove?: number;
-	// minimum from para to top of article
+	/**
+	 * Minimum distance from paragraph to top of article
+	 */
 	minAbove: number;
-	// minimum from (top of) para to bottom of article
+	/**
+	 * Minimum distance from (top of) paragraph to bottom of article
+	 */
 	minBelow: number;
-	// vertical px to clear the content meta element (byline etc) by. 0 to ignore
-	// used for carrot ads
+	/**
+	 * Vertical px to clear the content meta element (byline etc) by. 0 to ignore.
+	 * used for carrot ads
+	 */
 	clearContentMeta?: number;
-	// This is a map of selectors to rules. Each selector will be used to find opponents which are elements that we want to avoid placing ads too close to. If the opponent is too close to a candidate by the specified minAboveSlot or minBelowSlot, the candidate will be excluded.
+	/**
+	 * This is a map of selectors to rules. Each selector will be used to find opponents
+	 * which are elements that we want to avoid placing ads too close to. If the opponent
+	 * is too close to a candidate by the specified minAboveSlot or minBelowSlot, the
+	 * candidate will be excluded.
+	 */
 	opponentSelectorRules?: Record<string, RuleSpacing>;
-	// will run each slot through this fn to check if it must be counted in
+	/**
+	 * Will run each slot through this fn to check if it must be counted in
+	 */
 	filter?: (x: SpacefinderItem, lastWinner?: SpacefinderItem) => boolean;
-	// will remove slots before this one
+	/**
+	 * Will remove slots before this one
+	 */
 	startAt?: HTMLElement;
-	// will remove slots from this one on
+	/**
+	 * Will remove slots from this one on
+	 */
 	stopAt?: HTMLElement;
-	// will reverse the order of slots (this is useful for lazy loaded content)
+	/**
+	 * Will reverse the order of slots (this is useful for lazy loaded content)
+	 */
 	fromBottom?: boolean;
 };
 
@@ -210,7 +239,9 @@ const partitionCandidates = <T>(
  *
  * The candidate is the element where we would like to insert an ad above. Candidates satisfy the `selector` rule.
  *
- * Opponents are other elements in the article that are in the spacefinder ruleset for the current pass. This includes slots inserted by a previous pass but not those in the current pass as they're all inserted at the end.
+ * Opponents are other elements in the article that are in the spacefinder ruleset
+ * for the current pass. This includes slots inserted by a previous pass but not
+ * those in the current pass as they're all inserted at the end.
  *
  *                                                        │
  *                     Opponent Below                     │             Opponent Above